requirements.txt
google.generativeai
requests
rich
